refactor: application 폴더 분리에 따른 import 수정
