+++ src/hooks/admin/ApplicationClosePage/useClose.ts
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { convertPath } from 'apis/convertURI';
-import { getRecommends, postRecommends } from 'apis/admin/application';
 import { useQuery } from '@tanstack/react-query';
+import { getRecommends, postRecommends } from 'apis/admin/application/close';
 
 const useClose = (startWeekDate: string) => {
   // 후보 별 weekly 스케줄 정보 불러오기
@@ -21,7 +21,7 @@ const useClose = (startWeekDate: string) => {
   // 제출 클릭시 post 요청
   const navigate = useNavigate();
   const submitHandler = () => {
-    postRecommends({ selection: candidate })
+    postRecommends({ selection: candidate + 1 })
       .then((res) => {
         navigate(convertPath('/'));
       })
