+++ src/pages/SelectWeekPage/AdminDetailSection/InProgressDetail.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useQuery } from '@tanstack/react-query';
-import { getApplyStatus } from 'apis/admin/application';
+import { getApplyStatus } from 'apis/admin/application/status';
 import { convertPath } from 'apis/convertURI';
 import SubmitButton from 'components/@commons/SubmitButton';
 import DailyWorkersTemplate from 'components/DailyWorkers/DailyWorkersTemplate';
