+++ src/pages/admin/ApplicationOpenPage/index.tsx
@@ -4,7 +4,7 @@ import { atom, useAtomValue } from 'jotai';
 import PageContainer from 'components/@commons/PageContainer';
 import SetPeopleSection from './SetPeopleSection';
 import SetTimeSection from './SetTimeTemplateSection';
-import { TimeData } from 'apis/admin/application';
+import { TimeData } from 'apis/admin/application/open';
 
 export const timeTemplateAtom = atom<TimeData[]>([]);
 export const weeklyPeopleAtom = atom<number[][]>([[], [], [], [], [], [], []]);
